+++ src/main/index.ts
@@ -1,10 +1,21 @@
 console.time('init')
 
 import { mkdirSync, Stats, watchFile } from 'fs'
-import { app as rawApp, dialog } from 'electron'
+import { app as rawApp, dialog, protocol } from 'electron'
 import rc from './rc'
 import { VERSION, GIT_REF, BUILD_TIMESTAMP } from '../shared/build-info'
 import type { EventEmitter } from 'events'
+
+protocol.registerSchemesAsPrivileged([
+  {
+    scheme: 'webxdc',
+    privileges: {
+      allowServiceWorkers: true,
+      standard: true,
+      supportFetchAPI: true,
+    },
+  },
+])
 
 const app = rawApp as ExtendedAppMainProcess
 app.rc = rc
@@ -185,6 +196,13 @@ app.on('web-contents-created', (_e, contents) => {
   contents.setWindowOpenHandler(_details => {
     // prevent new windows from being created when clicking on links
     return { action: 'deny' }
+  })
+
+  // Prevent webview tags from being created,
+  // if you need them make sure to read https://www.electronjs.org/docs/latest/tutorial/security#12-verify-webview-options-before-creation
+  // to not indroduce security risks
+  contents.on('will-attach-webview', (event, _webPreferences, _params) => {
+    event.preventDefault()
   })
 })